+++ tests/conftest.py
@@ -1,0 +1,2 @@
+# https://docs.pytest.org/en/latest/reference.html#testdir
+pytest_plugins = "pytester"