+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased - 2019-10-13
+
+- repka.api.json_.DictJsonRepo - repository for working with json data
+
 ## 0.6.0 - 2019-10-04
 
 ### Changed